+++ bittytax/conv/parsers/bscscan.py
@@ -12,6 +12,10 @@ def parse_bscscan(data_row, _parser, **_kwargs):
     row_dict = data_row.row_dict
     data_row.timestamp = DataParser.parse_timestamp(int(row_dict['UnixTimestamp']))
 
+    if row_dict['Status'] != '':
+        # Failed txns should not have a Value_OUT
+        row_dict['Value_OUT(BNB)'] = 0
+
     if Decimal(row_dict['Value_IN(BNB)']) > 0:
         data_row.t_record = TransactionOutRecord(TransactionOutRecord.TYPE_DEPOSIT,
                                                  data_row.timestamp,
@@ -75,6 +79,22 @@ DataParser(DataParser.TYPE_EXPLORER,
            ['Txhash', 'Blockno', 'UnixTimestamp', 'DateTime', 'From', 'To', 'ContractAddress',
             'Value_IN(BNB)', 'Value_OUT(BNB)', None, 'TxnFee(BNB)', 'TxnFee(USD)',
             'Historical $Price/BNB', 'Status', 'ErrCode', 'PrivateNote'],
+           worksheet_name="BscScan",
+           row_handler=parse_bscscan)
+
+DataParser(DataParser.TYPE_EXPLORER,
+           "BscScan (BSC Transactions)",
+           ['Txhash', 'Blockno', 'UnixTimestamp', 'DateTime', 'From', 'To', 'ContractAddress',
+            'Value_IN(BNB)', 'Value_OUT(BNB)', None, 'TxnFee(BNB)', 'TxnFee(USD)',
+            'Historical $Price/BNB', 'Status', 'ErrCode', 'Method'],
+           worksheet_name="BscScan",
+           row_handler=parse_bscscan)
+
+DataParser(DataParser.TYPE_EXPLORER,
+           "BscScan (BSC Transactions)",
+           ['Txhash', 'Blockno', 'UnixTimestamp', 'DateTime', 'From', 'To', 'ContractAddress',
+            'Value_IN(BNB)', 'Value_OUT(BNB)', None, 'TxnFee(BNB)', 'TxnFee(USD)',
+            'Historical $Price/BNB', 'Status', 'ErrCode', 'Method', 'PrivateNote'],
            worksheet_name="BscScan",
            row_handler=parse_bscscan)